requirements.txt
appnope==0.1.4
astroid==3.2.2
asttokens==2.4.1
attrs==23.2.0
autopep8==2.3.1
beautifulsoup4==4.11.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
comm==0.2.2
configparser==7.0.0
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
dill==0.3.8
et_xmlfile==2.0.0
executing==2.0.1
fonttools==4.54.1
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
isort==5.13.2
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.2.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
msal==1.30.0
nest-asyncio==1.6.0
numpy==2.0.0
oauth2client==4.1.3
oauthlib==3.2.2
Office365-REST-Python-Client==2.5.12
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
PyDrive==1.3.1
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.2.5
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
seaborn==0.13.2
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tomlkit==0.12.5
tornado==6.4.1
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.4
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
